Phase-2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "openai-chatkit",
    "openai-agents",
    "fastapi[standard]>=0.124.4",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pyjwt==2.8.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.27",
    "python-dotenv",
    "requests",
    "httpx",
]
